+++ esniper.c
@@ -836,7 +836,7 @@ main(int argc, char *argv[])
 		numAuctions = readAuctionFile(options.auctfilename, &auctions);
 	} else {
 		numAuctions = argc / 2;
-		auctions = (auctionInfo **)myMalloc(numAuctions * sizeof(auctionInfo *));
+		auctions = (auctionInfo **)myMalloc((size_t)numAuctions * sizeof(auctionInfo *));
 		for (i = 0; i < argc/2; i++)
 			auctions[i] = newAuctionInfo(argv[2*i], argv[2*i+1]);
 	}
